+++ CHANGELOG.md
@@ -52,6 +52,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [#46](https://github.com/WSH032/pytauri/pull/46) - bump `tauri` to `v2.2`
 
+### Internal
+
+- [#64](https://github.com/WSH032/pytauri/pull/64) - test: add integration tests for `command` and `channel` ipc
+
 ## [0.1.0-beta]
 
 [unreleased]: https://github.com/WSH032/pytauri/tree/HEAD